+++ src/plugins/apollo.js
@@ -5,8 +5,7 @@ import VueApollo from 'vue-apollo'
 
 export default ({ app, Vue }) => {
   const httpLink = new HttpLink({
-    uri: 'http://localhost:5000/graphql', // PostGraphile
-//  uri: 'http://localhost:8090/v1alpha1/graphql', // Hasura
+    uri: 'http://localhost:5000/graphql'
   })
 
   // Create the apollo client