+++ ChangeLog.md
@@ -1,4 +1,12 @@
+2018-10-15
+
+- dockerize pxs-ui. Now, all components can be launched via docker-compose
+
+
 2018-10-12
+
+- model renaming: from "task" to "mission"
+- model simplification: remove "plan"
 
 - task dispatch:
     - has an "argument" table with all parameters